finished mod02

+++ module_02/ex02/Fixed.cpp
@@ -6,7 +6,7 @@
 /*   By: aalcara- <aalcara-@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/15 10:32:45 by aalcara-          #+#    #+#             */
-/*   Updated: 2022/02/27 16:24:40 by aalcara-         ###   ########.fr       */
+/*   Updated: 2022/02/28 11:06:15 by aalcara-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -204,7 +204,7 @@ Fixed &Fixed::operator-- ()
 
 const Fixed &Fixed::min (const Fixed &Obj1, const Fixed &Obj2)
 {
-	if (Obj1.toFloat() < Obj2.toFloat())
+	if (Obj1 < Obj2)
 		return (Obj1);
 	else
 		return (Obj2);
@@ -212,7 +212,7 @@ const Fixed &Fixed::min (const Fixed &Obj1, const Fixed &Obj2)
 
 Fixed &Fixed::min (Fixed &Obj1, Fixed &Obj2)
 {
-	if (Obj1.toFloat() < Obj2.toFloat())
+	if (Obj1 < Obj2)
 		return (Obj1);
 	else
 		return (Obj2);
@@ -220,7 +220,7 @@ Fixed &Fixed::min (Fixed &Obj1, Fixed &Obj2)
 
 const Fixed &Fixed::max (const Fixed &Obj1, const Fixed &Obj2)
 {
-	if (Obj1.toFloat() > Obj2.toFloat())
+	if (Obj1 > Obj2)
 		return (Obj1);
 	else
 		return (Obj2);
@@ -228,7 +228,7 @@ const Fixed &Fixed::max (const Fixed &Obj1, const Fixed &Obj2)
 
 Fixed &Fixed::max (Fixed &Obj1, Fixed &Obj2)
 {
-	if (Obj1.toFloat() > Obj2.toFloat())
+	if (Obj1 > Obj2)
 		return (Obj1);
 	else
 		return (Obj2);
